refactor(analytics-docs): Simplify App component and remove unused files
