+++ packages/analytics-docs/src/utils/getPlatformIcon.ts
